+++ src/types/withdrawal.ts
@@ -48,7 +48,7 @@ export interface ERC20PrepareWithdrawal {
 export type TokenWithdrawal = ETHWithdrawal | ERC20Withdrawal | ERC721Withdrawal;
 export type TokenPrepareWithdrawal = ETHPrepareWithdrawal | ERC20PrepareWithdrawal | ERC721Withdrawal;
 
-export type PrepareWithdrawalRequest = WalletConnection & {
+export type PrepareWithdrawalRequest =  {
   token: TokenPrepareWithdrawal;
   quantity: string;
 }